balance: item weight rebalance — trivial items 0.1~0.5, gear 14~15
负重系统改为 double 类型（支持小数重量）

重量调整：
  纸条/酒杯    1~2 → 0.1      金币       10 → 0.5
  钥匙×2       1~2 → 0.5      水晶碎片    3 → 0.5
  草药         2   → 0.5      传送粉      2 → 0.5
  火把         3   → 1        魔法饼干    1 → 0.5
  光之宝石     5   → 3        古书       15 → 6
  旧木桶      20   → 8
  锈剑        25   → 15       木盾       18 → 14

全部物品总重: 110 → 50.2 (刚好装满基础负重)
吃魔法饼干后上限 70，全捡无压力

基础设施: Item/Player/GameState/Room 的 weight 统一改为 double，
UiDrawUtils.drawStatusBar 支持小数显示

+++ src/main/java/cn/edu/whut/sept/zuul/engine/ItemManager.java
@@ -95,7 +95,7 @@ public class ItemManager
     {
         Item item = engine.getPlayer().removeItem(itemId);
         if (item != null && item.isMagicCookie()) {
-            int before = engine.getPlayer().getMaxWeight();
+            double before = engine.getPlayer().getMaxWeight();
             engine.getPlayer().setMaxWeight(before + Item.COOKIE_WEIGHT_BOOST);
         } else if (item != null) {
             engine.getPlayer().addItem(item);
